переделка прогресс бар через useState
